+++ python/CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to the AIStore Python SDK project are documented in this fil
 
 We structure this changelog in accordance with [Keep a Changelog](https://keepachangelog.com/) guidelines, and this project follows [Semantic Versioning](https://semver.org/).
 
----
 ## Unreleased
 
 ### Added
@@ -17,6 +16,7 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 
 - Make usage clear for ObjectReader API in `Object.get` deprecation message.
 - Make usage clear for ObjectWriter API in `Object.put_content`, `Object.put_file`, `Object.append_content`, and `Object.set_custom_props` deprecation messages.
+- Add `list_archive` function to Bucket class: helper method to list entries inside an archived object, with the option to include the archive itself.
 
 ## [1.14.0] - 2025-05-27